Factor operations dict out of `ship_computer` function
This commit factors out the creation of partial functions for the
ship computer's addition and multiplication operations, as it's not
useful to recreate them every time the computer runs.

In addition, I've also changed the name of the function `_math` to
something more generic to signify that it's a generic CPU operation.