Add solution to day 7 of the Advent of Code
For this puzzle, I decided to finally create a special IntCode helper
as a lot of puzzles seem to involve the IntCode Computer. The current
implementation can be run in multiple threads and communicates over
std.in. adn std.out. pipes represented by `Queue.SimpleQueue`.

Using this, day 7's puzzle was fairly straightforwards. I used a
`ThreadPool` to run the different "amps" (intcode threads) and a
`ProcessPool` to quickly run all possible configurations.