Factor out get_data function from timeit
The `get_data` function was part of the `timeit` loop that kept track
of the execution time of the solutions. Since this utility function
potentially makes a lengthy API call and carries out tasks that are
not strictly necessary, this skewed the results.

The sulution was to get the data prior to calling `default_timer`.