Add support for relative pointers to IntCodeApplication
As day 9 specified that our IntCode processor needed to support
relative pointers, I have added support for them to the class by
creating a new `write` method that takes care of writing the values
and adding a relative mode to the `read` method.

I have also cleaned up some parts of the code.